current_code/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
nova-act>=0.1.0
google-generativeai>=0.3.0
